AMB roles/MedisDePago/Insumos Fix en tipo de datos de insumos, correccion de stringconexion

+++ Proyecto Integrador/Entidades/Insumos.cs
@@ -8,11 +8,11 @@ namespace Entidades
 {
     public class Insumos
     {
-   public int Id_insumo { get; set; }   
-  public string Sku { get; set; }
-   public string Nombre { get; set; }
-   public float Precio { get; set; }
-   public int Cantidad { get; set; }
+        public int Id_insumo { get; set; }
+        public string Sku { get; set; }
+        public string Nombre { get; set; }
+        public double Precio { get; set; }
+        public int Cantidad { get; set; }
 
     }
 }
